test(kubernetes): bootstrapped integration tests
Integration tests for the mcp-server-kubernetes running with a MockWebServer.

These tests should be able to run in any OS (as opposed to those based on TestContainers).

These tests rely on langchain4j mcp to create an stdio tranpsport client.

Tests can be run in JVM and native profile.
Passing locally on an x86_64 linux machine.